requirements.txt
fastapi==0.73.0
uvicorn==0.17.4
motor==2.5.1
pydantic[email]
python-multipart
pillow